backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.2.0
openai==1.35.10
pydantic==2.8.2    